+++ zeroeq/connection/broker.cpp
@@ -1,5 +1,5 @@
 
-/* Copyright (c) 2014-2015, Human Brain Project
+/* Copyright (c) 2014-2017, Human Brain Project
  *                          Stefan.Eilemann@epfl.ch
  */